Support for prefixed conda environments
Addresses #26

When working with many environments it's easy to loose track of all the
environments. One handy solution for this is to have a prefixed environment in
each project's working directory. This commit adds support for these path
prefixed conda environments.

This change also works for environments created by virtualenv.

Details of change:

* Modify `conda-env-stripped-path` so that when a path prefixed environment is
  deactivated, `conda-env-current-name` is the path to be stripped. So no
  expanseion to `conda-env-default-location` etc. is appropriate.

* `conda-env-deactivate` is only doing something if an environment is
  actually activated.

* Introduce `conda-env-is-valid-path`

* Introduce interactive `conda-env-activate-path` which is a modified version
  of `conda-env-activate`.