Refactor current-name -> current-path
The `conda-env-current-name` was being used for storing both path and
name of environments, depending on whether it was being called by
`conda-env-activate` or `conda-env-activate-path`.

This commit removes this ambiguity, and also removes duplicate code that
become unnecessary after the simplication.

Moreover, this commit fixes an existing bug where the package was unable
to deactivate a base environment set with the hack
`(conda-env-activate "..")`.