Fix #68 `conda-env-activate-path` for Windows
The function `conda-env-activate-path` was using the function
`conda-env-is-valid-path` to check the validity of a conda env. This
function was probably never compatible with Windows since for those
systems `python.exe` lays on the root of the env dir, as in
https://docs.anaconda.com/anaconda/user-guide/tasks/integration/python-path/

This commit fixes this error by using an existing function to check the
validity of environments, namely, `conda--env-dir-is-valid`.