Fix #69 conda-env-current-name undefined
The variable `conda-env-current-name` has been removed in a recent
refactoring due to it being no longer necessary for operational purposes
of the package. Also, it had an ambiguous use and thus the reason for
the previous removal, see 5dc78b5.

However, now I realize it is important for having a nicer interface when
activating a conda environment, running `conda-env-shell-init`,
displaying env name in the modeline. This commit restores this variable
and associated functionality.