Register process sentinel to log status on process exit
Enables lexical scoping, which is needed for closures to work in sentinel callbacks.

The complexity in this handler is due to the need to ensure that sentinels
registered upfront a preserved.

For example I pronally have a hook registered for term-mode, which registers
a sentinel to prompt for closing the buffer on process exit. Without this logic
this completion status logger would override my handler completely.