fixed bug with conda-env-manage-for-buffer
When called with only one prefix argument, while no environment YAML file is found.

In such case it will conda-env-yaml-process-for-buffer will try to call conda create -f nil,
which will throw an error because it expects string as filename.

Instead, the behaviour should be fallback to generate a new environment file.